Protect routes with user auth

+++ controllers/movieController.js
@@ -1,24 +1,27 @@
 const asyncHandler = require('express-async-handler')
 const Movie = require('../models/Movie')
+const User = require('../models/User')
 
 // Get all movies
 // GET /api/movies
 const getMovies = asyncHandler (async (req, res) => {
-    const movies = await Movie.find()
+    const movies = await Movie.find({user: req.user.id})
     res.status(200).json(movies)
 })
 
 // Create a movie
 // POST /api/movies
 const createtMovie = asyncHandler (async (req, res) => {
-    if (!req.body.title) {
+    const {title, director, synopsis} = req.body
+    if (!title || !director || !synopsis) {
         res.status(400)
-        throw new Error('Please add a text field')
+        throw new Error('All fields are required')
     }
     const movie = await Movie.create({
         title: req.body.title,
         director: req.body.director,
-        synopsis: req.body.synopsis
+        synopsis: req.body.synopsis,
+        user: req.user.id
 
     })
     res.status(200).json(movie)
@@ -32,6 +35,15 @@ const updateMovie = asyncHandler (async (req, res) => {
         res.status(400)
         throw new Error('Movie not found')
     }
+    const user = await User.findById(req.user.id)
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    if(movie.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
+    }
     const updatedMovie = await Movie.findByIdAndUpdate(req.params.id, req.body, {new: true})
     res.status(200).json(updatedMovie)
 })
@@ -43,6 +55,15 @@ const deleteMovie = asyncHandler (async (req, res) => {
     if (!movie) {
         res.status(400)
         throw new Error('Movie not found')
+    }
+    const user = await User.findById(req.user.id)
+    if(!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+    if(movie.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorized')
     }
     await movie.deleteOne()
     res.status(200).json({id: req.params.id})
